+++ internal/machines/gc.go
@@ -19,7 +19,7 @@ type bucket struct {
 	samples int // Number of datasets to keep / bucket
 }
 
-const timeDay = 24 * int(time.Hour)
+const timeDay = 24 * int64(time.Hour)
 const timeFormat = "2006-01-02 15:04:05"
 
 type sortedReverseByTimeStates []*State
